+++ src/main/java/com/zipcodewilmington/heroku_demo/CityDAO.java
@@ -4,5 +4,5 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CityDAO extends PagingAndSortingRepository<Long, City> {
+public interface CityDAO extends PagingAndSortingRepository<Integer, City> {
 }
